Small documentation update with some packet corrections

+++ Sources/XTerraPacket/Types/PacketPlayerAddBuff.swift
@@ -3,7 +3,7 @@
 //
 //
 //  Created by Quentin Berry on 5/7/20.
-//
+//  Direction: Server <-> Client (Sync)
 
 import Foundation
 import SwiftyBytes
@@ -14,7 +14,7 @@ public struct PacketPlayerAddBuff: TerrariaPacket{
     public var packetType: TerrariaPacketType = .PlayerAddBuff
     public var payload: [UInt8] = []
     public var playerId: UInt8 = 0
-    public var buff: UInt8 = 0
+    public var buff: UInt16 = 0
     public var time: Int32 = 0
     
     public init(){}
@@ -26,14 +26,14 @@ public struct PacketPlayerAddBuff: TerrariaPacket{
         let data = BinaryData(data: self.payload)
         let reader = BinaryReader(data)
         self.playerId = try reader.readUInt8()
-        self.buff = try reader.readUInt8()
+        self.buff = try reader.readUInt16()
         self.time = try reader.readInt32()
     }
     mutating public func encodePayload() throws{
         self.resetPayload()
         let writer = BinaryWriter()
         try writer.writeUInt8(playerId)
-        try writer.writeUInt8(buff)
+        try writer.writeUInt16(buff)
         try writer.writeInt32(time)
         payload.append(contentsOf: writer.data.bytes)
     }
